fixed a bug with first image upload crashing server, added hints to classic region creation

+++ app.js
@@ -69,7 +69,9 @@ var storage = multer.diskStorage({
 			cb(null,Date.now()+'.wav');
 		} else if (file.originalname.endsWith('.png')) {
 			var filename = req.user.username+'.png';
-			fs.unlinkSync('./uploads/img/'+filename);
+			if (fs.existsSync('./uploads/img/'+filename)) {
+				fs.unlinkSync('./uploads/img/'+filename);
+			}
 			cb(null, filename);
 		}
 	}
